# openapi.yaml
openapi: 3.0.3
info:
  title: Corgi Recommender Service API
  description: |
    API for the Corgi Recommender Service, a personalized recommendations system
    for Mastodon posts. The API handles user interactions, preferences, and generates
    personalized timelines.
    
    ## Key Features
    - **User Interaction Tracking**: Records favorites, bookmarks, and explicit feedback
    - **Personalized Recommendations**: Generates tailored post suggestions based on user activity
    - **Privacy Controls**: Provides user-configurable privacy settings
    - **Mastodon Compatibility**: Returns data in Mastodon-compatible format for easy integration
    - **Transparent Proxy**: Proxies requests to Mastodon instances with added recommendations
  version: 1.0.0
  contact:
    name: Corgi Service Team
    email: support@corgi-recommender.example.com
servers:
  - url: http://localhost:5001
    description: Local development server
  - url: https://api.corgi-recommender.example.com
    description: Production server
...
paths:
  /api/v1/interactions:
    post:
      summary: Log a user interaction with a post
      description: |
        Records user actions like favorites, bookmarks, or post feedback.
        This data is used to improve future recommendations.
      operationId: logInteraction
      tags:
        - interactions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InteractionRequest'
            example:
              user_alias: "user_12345"
              post_id: "post_67890"
              action_type: "favorite"
              context:
                source: "timeline_home"
      responses:
        '200':
          description: Interaction logged successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  interaction_id:
                    type: string
                    description: The ID of the logged interaction
                    example: "interaction_12345"
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/interactions/{post_id}:
    get:
      summary: Get interactions for a specific post
      description: Retrieves all recorded interactions for a single post
      operationId: getPostInteractions
      tags:
        - interactions
      parameters:
        - name: post_id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the post to retrieve interactions for
      responses:
        '200':
          description: Interactions for the post
          content:
            application/json:
              schema:
                type: object
                properties:
                  post_id:
                    type: string
                  interaction_counts:
                    $ref: '#/components/schemas/InteractionCounts'
                  interactions:
                    type: array
                    items:
                      type: object
                      properties:
                        action_type:
                          type: string
                        count:
                          type: integer
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /api/v1/timelines/home:
    get:
      summary: Get user's home timeline
      description: |
        Returns the user's home timeline, similar to a Mastodon home timeline.
        For real users, the request is proxied to their Mastodon instance.
        For test/synthetic users, returns mock data.
      operationId: getHomeTimeline
      tags:
        - timelines
        - proxy
      parameters:
        - name: user_id
          in: query
          required: false
          schema:
            type: string
          description: The user ID to get timeline for (for testing/validation)
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
          description: Maximum number of posts to return
        - name: strategy
          in: query
          required: false
          schema:
            type: string
            enum: [uniform, after_n, first_only, tag_match]
          description: Injection strategy to use for recommended posts
        - name: inject
          in: query
          required: false
          schema:
            type: boolean
            default: true
          description: Whether to inject recommended posts into the timeline
      responses:
        '200':
          description: Home timeline posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MastodonPost'
                description: Array of Mastodon-compatible posts, possibly with injected recommendations
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/timelines/home/augmented:
    get:
      summary: Get augmented home timeline with recommendations
      description: |
        Returns the user's home timeline with personalized recommendations injected.
        Can blend real Mastodon posts with recommended posts based on user preferences.
      operationId: getAugmentedTimeline
      tags:
        - timelines
        - recommendations
        - proxy
      parameters:
        - name: user_id
          in: query
          required: false
          schema:
            type: string
          description: The user ID to get timeline for (for testing/validation)
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
          description: Maximum number of posts to return
        - name: inject_recommendations
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: Whether to inject personalized recommendations into the timeline
      responses:
        '200':
          description: Augmented timeline posts
          content:
            application/json:
              schema:
                type: object
                properties:
                  timeline:
                    type: array
                    items:
                      $ref: '#/components/schemas/MastodonPost'
                  injected_count:
                    type: integer
                    description: Number of recommendations injected
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /api/v1/privacy:
    get:
      summary: Get user privacy settings
      description: Retrieves the current privacy settings for a user
      operationId: getPrivacySettings
      tags:
        - privacy
      parameters:
        - name: user_id
          in: query
          required: true
          schema:
            type: string
          description: The user ID to get privacy settings for
      responses:
        '200':
          description: User privacy settings
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: string
                  tracking_level:
                    $ref: '#/components/schemas/TrackingLevel'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      summary: Update user privacy settings
      description: |
        Updates a user's privacy settings, controlling how much data is collected
        and how it's used for recommendations.
      operationId: updatePrivacySettings
      tags:
        - privacy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
                - tracking_level
              properties:
                user_id:
                  type: string
                  description: The user ID to update privacy settings for
                tracking_level:
                  $ref: '#/components/schemas/TrackingLevel'
            example:
              user_id: "user_12345"
              tracking_level: "limited"
      responses:
        '200':
          description: Privacy settings updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: string
                  tracking_level:
                    $ref: '#/components/schemas/TrackingLevel'
                  status:
                    type: string
                    example: ok
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  schemas:
    InteractionRequest:
      type: object
      required:
        - post_id
        - action_type
      properties:
        user_alias:
          type: string
          description: Pseudonymized identifier for the user
        user_id:
          type: string
          description: Alternative direct user identifier
        post_id:
          type: string
          description: Unique identifier for the post
        action_type:
          type: string
          description: The type of interaction
          enum:
            - favorite
            - bookmark
            - reblog
            - more_like_this
            - less_like_this
        context:
          type: object
          description: Additional context about the interaction
          properties:
            source:
              type: string
              description: Source of the interaction (e.g., timeline_home)
...
    InteractionCounts:
      type: object
      properties:
        favorites:
          type: integer
          default: 0
        reblogs:
          type: integer
          default: 0
        replies:
          type: integer
          default: 0
        bookmarks:
          type: integer
          default: 0
    MastodonPost:
      type: object
      required:
        - id
        - content
        - created_at
        - account
      properties:
        id:
          type: string
          description: Unique identifier for the post
        content:
          type: string
          description: HTML content of the post
        created_at:
          type: string
          format: date-time
          description: Creation timestamp in ISO 8601 format
        language:
          type: string
          description: ISO language code (e.g., "en")
        account:
          type: object
          required:
            - id
            - username
            - display_name
          properties:
            id:
              type: string
              description: Unique identifier for the account
            username:
              type: string
              description: Username of the account
            display_name:
              type: string
              description: Display name of the account
            followers_count:
              type: integer
              description: Number of followers
            following_count:
              type: integer
              description: Number of accounts following
            statuses_count:
              type: integer
              description: Number of posted statuses
            url:
              type: string
              format: uri
              description: URL to the account's profile
        replies_count:
          type: integer
          description: Number of replies to this post
        reblogs_count:
          type: integer
          description: Number of reblogs of this post
        favourites_count:
          type: integer
          description: Number of favorites for this post
        url:
          type: string
          format: uri
          description: URL to the post
        sensitive:
          type: boolean
          description: Whether the post contains sensitive content
        spoiler_text:
          type: string
          description: Content warning text
        visibility:
          type: string
          enum: [public, unlisted, private, direct]
          description: Visibility level of the post
        ranking_score:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: Recommendation ranking score (0.0 to 1.0)
        recommendation_reason:
          type: string
          description: Human-readable reason for recommendation
        is_recommendation:
          type: boolean
          description: Whether this post was added as a recommendation
        is_real_mastodon_post:
          type: boolean
          description: Whether this is a real Mastodon post (vs. synthetic)
        is_synthetic:
          type: boolean
          description: Whether this is a synthetic post
        injected:
          type: boolean
          description: Whether this post was injected by the recommendation system
        injection_metadata:
          type: object
          description: Metadata about the injected recommendation post
          properties:
            source:
              type: string
              description: Source of the recommendation (e.g., recommendation_engine, timeline_injector)
              example: recommendation_engine
            strategy:
              type: string
              description: Strategy used for the recommendation (e.g., personalized, tag_match)
              example: personalized
            explanation:
              type: string
              description: Human-readable explanation for why this post was recommended
              example: Suggested based on your recent interests in #python and #opensource
            score:
              type: number
              format: float
              description: Recommendation confidence score (0-1)
              example: 0.87
...
    TrackingLevel:
      type: string
      description: |
        Privacy level controlling data collection and usage:
        - full: All interaction data is collected and used for personalization
        - limited: Only aggregated data is stored, with limited personalization
        - none: No tracking of interactions, only basic functionality
      enum:
        - full
        - limited
        - none
  responses:
    BadRequest:
      description: Bad request due to invalid parameters or payload
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              received:
                type: object
                additionalProperties: true
    NotFound:
      description: The requested resource was not found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              type:
                type: string
              request_id:
                type: string
...